requirements.txt

Flask==2.2.2
Werkzeug==2.2.3
psycopg2-binary==2.9.5
